map store name to id
